PptxTranslator.Web/requirements.txt
python-pptx
deep-translator
